+++ infrastructures/api/server.go
@@ -1,4 +1,4 @@
-package infrastructure
+package api
 
 import (
 	"net/http"
@@ -9,19 +9,17 @@ import (
 )
 
 // NewRouter create a new instance of http server and route.
-func NewRouter() {
+func NewRouter(todoController *controllers.TodoController) {
 	e := echo.New()
 
 	e.Use(middleware.Logger())
 	e.Use(middleware.Recover())
 
-	todoController := controllers.NewTodoController(NewSQLHandler())
-
 	e.GET("/", func(c echo.Context) error {
 		return c.String(http.StatusOK, "Hello, World!")
 	})
 	e.GET("/todos", func(c echo.Context) error {
-		return todoController.Show(c)
+		return todoController.GetTodos(c)
 	})
 
 	e.Logger.Fatal(e.Start(":1323"))
