--- server.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	infrastructure "github.com/halkn/echo-sample/infrastructures"
-)
-
-func main() {
-	infrastructure.NewRouter()
-}
